+++ test/extension/index.js
@@ -13,6 +13,7 @@ exports.run = async function run() {
     'ideaGit.showLog',
     'ideaGit.openResourceDiff',
     'ideaGit.showFileHistory',
+    'ideaGit.showLineBlame',
     'ideaGit.moveResourceToChangelist',
     'ideaGit.showResourceInChanges',
     'ideaGit.changes.focus',